+++ matrix_operations.cpp
@@ -15,9 +15,9 @@ Tensor wef::matmul(const Tensor& m1, const Tensor& m2)
 
     Tensor m = Tensor::create(temp_shape.get(), m1.rank);
 
-    const double* pm1 = m1.tensor.get(); // grab raw pointers for speeeed
-    const double* pm2 = m2.tensor.get();
-    double* pm = m.tensor.get();
+    const float* pm1 = m1.tensor.get(); // grab raw pointers for speeeed
+    const float* pm2 = m2.tensor.get();
+    float* pm = m.tensor.get();
 
     const size_t m1size = m1.row * m1.col; // to shift pm1 by one batch worth
     const size_t m2size = m2.row * m2.col * not_bcast; // only shift if m2 is 3D
@@ -32,7 +32,7 @@ Tensor wef::matmul(const Tensor& m1, const Tensor& m2)
 
         size_t r = b * m1size + ((elem / m.col) % m.row) * m1.col;
 
-        double sum = 0;
+        float sum = 0;
         for (size_t i = 0; i < m1.col; i++)
                 sum += pm1[i + r] * pm2[i * m2.col + c];
                 
@@ -41,9 +41,9 @@ Tensor wef::matmul(const Tensor& m1, const Tensor& m2)
     */
 
     const size_t msize = m1.row * m2.col;
-    const double* pm1temp;
-    const double* pm2temp;
-    double* pmtemp;
+    const float* pm1temp;
+    const float* pm2temp;
+    float* pmtemp;
     for (int b = 0; b < m1.batch; b++){
         
         pm1temp = pm1 + b * m1size; // shift pm1 by one batch worth
@@ -54,7 +54,7 @@ Tensor wef::matmul(const Tensor& m1, const Tensor& m2)
         for (size_t i = 0; i < m1.row; i++) {
             for (size_t k = 0; k < m2.col; k++) {
 
-                double sum = 0;
+                float sum = 0.0f;
                 for (size_t j = 0; j < m1.col; j++) {
                     sum += pm1temp[i * m1.col + j] * pm2temp[j * m2.col + k];
                 }
@@ -79,16 +79,16 @@ Tensor wef::matmul(const Tensor& m1, const Tensor& m2, bool, int n_threads)
 
     Tensor m = Tensor::create(temp_shape.get(), m1.rank);
 
-    const double* pm1 = m1.tensor.get();
-    const double* pm2 = m2.tensor.get();
-    double* pm = m.tensor.get();
+    const float* pm1 = m1.tensor.get();
+    const float* pm2 = m2.tensor.get();
+    float* pm = m.tensor.get();
 
     const size_t m1size = m1.row * m1.col; // to shift pm1 by one batch worth
     const size_t m2size = m2.row * m2.col * not_bcast; // only shift if m2 is 3D
     const size_t msize = m1.row * m2.col;
-    const double* pm1temp;
-    const double* pm2temp;
-    double* pmtemp;
+    const float* pm1temp;
+    const float* pm2temp;
+    float* pmtemp;
 
     // multi thread additions
     if (n_threads == 0)
@@ -120,7 +120,7 @@ Tensor wef::matmul(const Tensor& m1, const Tensor& m2, bool, int n_threads)
                     for (size_t i = th * stride; i < (th * stride) + temp; i++) {
                         for (size_t k = 0; k < m2col; k++) {
 
-                            double sum = 0;
+                            float sum = 0;
                             for (size_t j = 0; j < m1col; j++) {
                                 sum += pm1temp[i * m1col + j] * pm2temp[j * m2col + k];
                             }
@@ -142,12 +142,12 @@ Tensor wef::matmul(const Tensor& m1, const Tensor& m2, bool, int n_threads)
     return m;
 }
 
-Tensor wef::cops(const Tensor& m1, const double con, double (*f)(double, double)) 
+Tensor wef::cops(const Tensor& m1, const float con, float (*f)(float, float)) 
 {
     Tensor m = Tensor::create(m1.shape.get(), m1.rank);
 
-    const double* pm1 = m1.tensor.get();
-    double* pm = m.tensor.get();
+    const float* pm1 = m1.tensor.get();
+    float* pm = m.tensor.get();
 
     #pragma omp parallel for schedule(static)
     for (size_t i = 0; i < m1.tot_size; i++) pm[i] = f(pm1[i], con);
@@ -166,12 +166,12 @@ Tensor wef::transpose(const Tensor& m1)
     Tensor m = Tensor::create(temp_shape.get(), m1.rank);
 
 
-    const double* pm1 = m1.tensor.get();
-    double* pm = m.tensor.get();
+    const float* pm1 = m1.tensor.get();
+    float* pm = m.tensor.get();
 
     const size_t msize = m1.row * m1.col;
-    const double* pm1temp;
-    double* pmtemp;
+    const float* pm1temp;
+    float* pmtemp;
 
     for (int b = 0; b < m1.batch; b++){
 
@@ -196,13 +196,13 @@ Tensor wef::argmax(const Tensor& m1)
     for (int i = 0; i < m1.rank - 1; i ++) temp_shape[i] = m1.shape[i];
     Tensor m = Tensor::create(temp_shape.get(), m1.rank - 1);
 
-    const double* pm1 = m1.tensor.get();
-    double* pm = m.tensor.get();
+    const float* pm1 = m1.tensor.get();
+    float* pm = m.tensor.get();
 
     #pragma omp parallel for schedule(static)
     for (size_t i = 0; i < m1.tot_size / m1.col; i++)
     {
-        double temp_val = -1e19;
+        float temp_val = -1e19f;
         for (size_t j = 0; j < m1.col; j++) 
         {
             if (pm1[i * m1.col + j] > temp_val) {pm[i] = j; temp_val = pm1[i * m1.col + j];}
@@ -215,13 +215,13 @@ Tensor wef::softmax(const Tensor& m1)
 {
     Tensor m = Tensor::create(m1.shape.get(), m1.rank);
 
-    const double* pm1 = m1.tensor.get();
-    double* pm = m.tensor.get();
+    const float* pm1 = m1.tensor.get();
+    float* pm = m.tensor.get();
 
     #pragma omp parallel for schedule(static)
     for (size_t i = 0; i < m1.tot_size / m1.col; i++)
     {
-        double sum = 1e-19;
+        float sum = 1e-19f;
         for (size_t j = 0; j < m1.col; j++) sum += std::exp(pm1[i * m1.col + j]);
         for (size_t j = 0; j < m1.col; j++) pm[i * m1.col + j] = std::exp(pm1[i * m1.col + j]) / sum;
     }
@@ -232,8 +232,8 @@ Tensor wef::activation(const Tensor& m1, const char ops)
 {
     Tensor m = Tensor::create(m1.shape.get(), m1.rank);
 
-    const double* pm1 = m1.tensor.get();
-    double* pm = m.tensor.get();
+    const float* pm1 = m1.tensor.get();
+    float* pm = m.tensor.get();
 
     #pragma omp parallel for schedule(static)
     for (size_t i = 0; i < m1.tot_size; i++) 
@@ -241,11 +241,11 @@ Tensor wef::activation(const Tensor& m1, const char ops)
         switch (ops) {
             // relu
             case 'a':
-                pm[i] = pm1[i] > 0 ?  pm1[i] : 0.0;
+                pm[i] = pm1[i] > 0 ?  pm1[i] : 0.0f;
                 break;
             // derivative relu
             case 'b':
-                pm[i] = pm1[i] > 0 ? 1.0 : 0.0;
+                pm[i] = pm1[i] > 0 ? 1.0f : 0.0f;
                 break;
             // sigmoid
             case 'c':
@@ -277,11 +277,11 @@ Tensor wef::reducesum(const Tensor& m1, const int ax)
     }
 
     Tensor m = Tensor::create(out_shape.get(), m1.rank); 
-    const double* pm1 = m1.tensor.get();
-    double* pm = m.tensor.get();
+    const float* pm1 = m1.tensor.get();
+    float* pm = m.tensor.get();
     
     const size_t m1size = m1.row * m1.col;
-    std::memset(pm, 0, (m.tot_size) * sizeof(double));
+    std::memset(pm, 0, (m.tot_size) * sizeof(float));
 
     int eaa = 1; // everything after axis i.e. b, h w, axis, x1, x2 -> eaa = x1 * x2
     for (int i = ax + 1; i < m1.rank; i++) eaa *= m1.shape[i];
@@ -293,14 +293,14 @@ Tensor wef::reducesum(const Tensor& m1, const int ax)
     return m;
 }
 
-double wef::l2(const Tensor& m1, const Tensor& m2)
+float wef::l2(const Tensor& m1, const Tensor& m2)
 {
     if (m1.rank != m2.rank) throw std::invalid_argument("matrix rank mismatch [6]");
     for (int i = 0; i < m1.rank; i++) if (m1.shape[i] != m2.shape[i]) throw std::invalid_argument("matrix size mismatch [7]");
 
-    const double* pm1 = m1.tensor.get(); // grab raw pointers for speeeed
-    const double* pm2 = m2.tensor.get();
-    double loss = 0.0;
+    const float* pm1 = m1.tensor.get(); // grab raw pointers for speeeed
+    const float* pm2 = m2.tensor.get();
+    float loss = 0.0f;
 
     #pragma omp parallel for reduction(+:loss) schedule(static) 
     for (size_t i = 0; i < m1.tot_size; i++) 
@@ -311,14 +311,14 @@ double wef::l2(const Tensor& m1, const Tensor& m2)
     return loss / (m1.tot_size);
 }
 
-double wef::binarycrossentropy(const Tensor& m1, const Tensor& m2) // m1 is real and m2 pred !!
+float wef::binarycrossentropy(const Tensor& m1, const Tensor& m2) // m1 is real and m2 pred !!
 {
     // TO DO: catch mismatch tensor
 
-    const double* pm1 = m1.tensor.get(); // grab raw pointers for speeeed
-    const double* pm2 = m2.tensor.get();
-    double loss = 0.0;
-    const double eps = 1e-19;
+    const float* pm1 = m1.tensor.get(); // grab raw pointers for speeeed
+    const float* pm2 = m2.tensor.get();
+    float loss = 0.0f;
+    const float eps = 1e-19f;
     
     #pragma omp parallel for reduction(+:loss) schedule(static) 
     for (size_t i = 0; i < m1.tot_size; i++)
@@ -331,18 +331,18 @@ double wef::binarycrossentropy(const Tensor& m1, const Tensor& m2) // m1 is real
     return loss / m1.tot_size;
 }
 
-double wef::categoricalcrossentropy(const Tensor& m1, const Tensor& m2, Tensor& m /*m is same as pred*/) // m1 is real and m2 pred !!
+float wef::categoricalcrossentropy(const Tensor& m1, const Tensor& m2, Tensor& m /*m is same as pred*/) // m1 is real and m2 pred !!
 {
     // Note m1 is actual labels and m2 is probabilities 
     // eg: m1 = {{1}, {2}}, m2 = {{0, 1, 0}, {0, 0, 1}}
 
     // TO DO: catch mismatch tensor
 
-    const double* pm1 = m1.tensor.get(); // grab raw pointers for speeeed
-    const double* pm2 = m2.tensor.get();
-    double* pm = m.tensor.get();
-    double loss = 0.0;
-    const double eps = 1e-19;
+    const float* pm1 = m1.tensor.get(); // grab raw pointers for speeeed
+    const float* pm2 = m2.tensor.get();
+    float* pm = m.tensor.get();
+    float loss = 0.0f;
+    const float eps = 1e-19f;
     
     const size_t num_classes = m2.shape[m2.rank - 1];
 
@@ -353,15 +353,15 @@ double wef::categoricalcrossentropy(const Tensor& m1, const Tensor& m2, Tensor& 
         size_t base = i * m.shape[m.rank - 1];
 
         // find max per class and subtract to make stable
-        double cur_max = pm2[tempid];
+        float cur_max = pm2[tempid];
         for (size_t j = 0; j < num_classes; j++) { if (pm2[tempid + j] > cur_max) cur_max = pm2[tempid + j]; }
     
-        double sum = 1e-19;
+        float sum = 1e-19f;
         for (size_t j = 0; j < num_classes; j++) sum += std::exp(pm2[tempid + j] - cur_max);
         
         for (size_t j = 0; j < num_classes; j++)
         {
-            double p = std::exp(pm2[tempid + j] - cur_max) / sum;
+            float p = std::exp(pm2[tempid + j] - cur_max) / sum;
             if (j == (size_t)pm1[i])
             {
                 loss -= std::log(p + eps);
@@ -375,17 +375,17 @@ double wef::categoricalcrossentropy(const Tensor& m1, const Tensor& m2, Tensor& 
     return loss / m1.tot_size;
 }
 
-double wef::categoricalcrossentropy(const Tensor& m1, const Tensor& m2) // m1 is real and m2 pred !!
+float wef::categoricalcrossentropy(const Tensor& m1, const Tensor& m2) // m1 is real and m2 pred !!
 {
     // TO DO: catch mismatch tensor
 
     // Note m1 is actual labels and m2 is probabilities 
     // eg: m1 = {{1}, {2}}, m2 = {{0, 1, 0}, {0, 0, 1}}
 
-    const double* pm1 = m1.tensor.get(); // grab raw pointers for speeeed
-    const double* pm2 = m2.tensor.get();
-    double loss = 0.0;
-    const double eps = 1e-19;
+    const float* pm1 = m1.tensor.get(); // grab raw pointers for speeeed
+    const float* pm2 = m2.tensor.get();
+    float loss = 0.0f;
+    const float eps = 1e-19f;
     
     const size_t num_classes = m2.shape[m2.rank - 1];
 
@@ -395,15 +395,15 @@ double wef::categoricalcrossentropy(const Tensor& m1, const Tensor& m2) // m1 is
         size_t tempid = i * num_classes;
 
         // find max per class and subtract to make stable
-        double cur_max = pm2[tempid];
+        float cur_max = pm2[tempid];
         for (size_t j = 0; j < num_classes; j++) { if (pm2[tempid + j] > cur_max) cur_max = pm2[tempid + j]; }
     
-        double sum = 1e-19;
+        float sum = 1e-19f;
         for (size_t j = 0; j < num_classes; j++) sum += std::exp(pm2[tempid + j] - cur_max);
         
         for (size_t j = 0; j < num_classes; j++)
         {
-            double p = std::exp(pm2[tempid + j] - cur_max) / sum;
+            float p = std::exp(pm2[tempid + j] - cur_max) / sum;
             if (j == (size_t)pm1[i]) loss -= std::log(p + eps);
         }
     }
@@ -447,7 +447,7 @@ void wef::print(const Tensor& m1, size_t* arr, size_t num, bool allc)
 
 }
 
-Tensor wef::pow(const Tensor& m1, const double con) { return cops(m1, con, [](double a, double b){ return std::pow(a, b); }); }
+Tensor wef::pow(const Tensor& m1, const float con) { return cops(m1, con, [](float a, float b){ return std::pow(a, b); }); }
 Tensor wef::relu(const Tensor& m1) { return activation(m1, 'a'); }
 Tensor wef::d_relu(const Tensor& m1) { return activation(m1, 'b'); }
 Tensor wef::sigmoid(const Tensor& m1) { return activation(m1, 'c'); }
